+++ library/Option.scala
@@ -1,6 +1,6 @@
 /* Copyright 2009-2014 EPFL, Lausanne */
 
-package leon.lang
+package leon
 
 import leon.annotation._
