+++ library/lang/package.scala
@@ -27,4 +27,38 @@ package object lang {
 
   @ignore
   def error[T](reason: String): T = sys.error(reason)
+    
+  /*sealed class TemplateCons(val postcond : Boolean) {
+    //TODO allow elegant ways of supporting templates with arbitrary number of arguments    
+    def template(templateFunc : Float => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Float,Float) => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Float,Float, Float)  => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Float,Float, Float, Float)  => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Float,Float, Float, Float, Float)  => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Float,Float, Float, Float, Float, Float)  => Boolean) : Boolean = postcond
+  }*/
+
+  @ignore
+  sealed class TemplateCons(val postcond : Boolean) {
+    //TODO allow elegant ways of supporting templates with arbitrary number of arguments    
+    def template(templateFunc : Int => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Int,Int) => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Int,Int, Int)  => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Int,Int, Int, Int)  => Boolean) : Boolean = postcond    
+    /*def template(templateFunc : (Float,Float, Float, Float, Float)  => Boolean) : Boolean = postcond    
+    def template(templateFunc : (Float,Float, Float, Float, Float, Float)  => Boolean) : Boolean = postcond*/
+  }
+  
+  @ignore
+  implicit def any2Template(postcond: Boolean): TemplateCons = {    
+    new TemplateCons(postcond)
+  }
+  
+  //a counter that counts the number of time steps
+  @ignore
+  def time: Int = throw new RuntimeException("Implementation not supported")
+  
+  //a counter that counts the depth
+  @ignore
+  def depth: Int = throw new RuntimeException("Implementation not supported")
 }
