--- library/src/main/scala/leon/Annotations.scala
@@ -1,12 +1,0 @@
-/* Copyright 2009-2013 EPFL, Lausanne */
-
-package leon
-
-import scala.annotation.StaticAnnotation
-
-object Annotations {
-  class induct extends StaticAnnotation
-  class axiomatize extends StaticAnnotation
-  class main extends StaticAnnotation
-  class monotonic extends StaticAnnotation
-}
