+++ testcases/ravi-testcases/numerical/SimpleLoop.scala
@@ -1,4 +1,4 @@
-import leon.Utils._
+import leon.lang._
 
 object SimpleLoop
 {
@@ -6,8 +6,8 @@ object SimpleLoop
 	  if(x < 0)
 	    0
 	  else 
-	    s(x-1) + 1
-	    
-	} ensuring(res => res != -1 template((a, b, c) => a*res + b*x + c <= 0))	
+	    s(x-1) + 1	   
+	} ensuring(res => res != -1 template((a,b,c) => a*res + b*x + c <= 0)) 
+	//ensuring(res => (res != -1) template((a, b, c) => a*res + b*x + c <= 0))	
 	//inductive generalization res >= 0	
 }
